Split out classes into their own files.
