+++ cpvc/UI/ListTree.cs
@@ -1,41 +1,11 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
-using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
-using System.Windows.Controls;
 
 namespace CPvC
 {
-    public class ListTreeNode<T>
-    {
-        public ListTreeNode(T historyEvent)
-        {
-            HistoryEvent = historyEvent;
-            Children = new List<ListTreeNode<T>>();
-        }
-
-        public ListTreeNode<T> Parent
-        {
-            get;
-            set;
-        }
-
-        public List<ListTreeNode<T>> Children
-        {
-            get;
-        }
-
-        public T HistoryEvent
-        {
-            get;
-            set;
-        }
-    }
-
     public abstract class ListTree<T>
     {
         public ListTree()
@@ -173,7 +143,7 @@ namespace CPvC
             if (oldIndex < newIndex)
             {
                 refreshIndex = oldIndex;
-                
+
                 while (count > 0)
                 {
                     ListTreeNode<T> temp = _horizontalNodes[oldIndex];
@@ -543,389 +513,4 @@ namespace CPvC
         //protected HashSet<T> _horizontalTies;
         protected HashSet<Tuple<T, T>> _verticalTies;
     }
-
-    public class HistoryListTree : ListTree<HistoryEvent>
-    {
-        public HistoryListTree(History history)
-        {
-            SetHistory(history);
-        }
-
-        //public ListTreeNode<HistoryEvent> Root
-        //{
-        //    get
-        //    {
-        //        return Root;
-        //    }
-
-        //    protected set
-        //    {
-        //        Root = value;
-        //    }
-        //}
-
-        //public System.Drawing.Point GetPosition(ListTreeNode<HistoryEvent> node)
-        //{
-        //    return GetPosition(node);
-        //}
-
-        public void SetHistory(History history)
-        {
-            if (_history != null)
-            {
-                _history.Auditors -= ProcessHistoryChange;
-            }
-
-            _history = history;
-
-            if (_history != null)
-            {
-                Init();
-                _history.Auditors += ProcessHistoryChange;
-            }
-        }
-
-        private void Init()
-        {
-            //_listTree = new ListTree<HistoryEvent>(_history.RootEvent);
-
-            InitRoot(_history.RootEvent);
-
-            List<HistoryEvent> nodes = new List<HistoryEvent>();
-            nodes.AddRange(_history.RootEvent.Children);
-
-            while (nodes.Any())
-            {
-                HistoryEvent historyEvent = nodes[0];
-                nodes.RemoveAt(0);
-
-                AddEventToListTree(historyEvent);
-
-                nodes.AddRange(historyEvent.Children);
-            }
-        }
-
-        public void ProcessHistoryChange(object sender, HistoryChangedEventArgs args)
-        {
-
-            NotifyPositionChangedEventArgs<HistoryEvent> changeArgs = UpdateListTree(args);
-            if (changeArgs != null)
-            {
-                PositionChanged?.Invoke(this, changeArgs);
-            }
-        }
-
-        static private bool InterestingEvent(HistoryEvent historyEvent)
-        {
-            if (historyEvent is RootHistoryEvent ||
-                historyEvent is BookmarkHistoryEvent ||
-                historyEvent.Children.Count != 1)
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        private ListTreeNode<HistoryEvent> Add(ListTreeNode<HistoryEvent> parent, HistoryEvent historyEvent)
-        {
-            ListTreeNode<HistoryEvent> child = new ListTreeNode<HistoryEvent>(historyEvent);
-
-            // Check first if child already exists in the tree! And that parent exists in the tree!
-
-            // Insert into children of parent first!
-            // But! We must either add a new child, or inject this child inbetween the parent and an existing child!
-            // Which case are we?
-
-            // This class should not be doing this! HistoryControl2 should be... ListTree should be made to be generic at some point!
-            int descendentChildIndex = 0;
-            while (descendentChildIndex < parent.Children.Count)
-            {
-                if (child.HistoryEvent.IsEqualToOrAncestorOf(parent.Children[descendentChildIndex].HistoryEvent))
-                {
-                    break;
-                }
-
-                descendentChildIndex++;
-            }
-
-            int childIndex = 0;
-            if (descendentChildIndex < parent.Children.Count)
-            {
-                ListTreeNode<HistoryEvent> descendentChild = parent.Children[descendentChildIndex];
-                childIndex = descendentChildIndex;
-
-                child.Children.Add(descendentChild);
-                descendentChild.Parent = child;
-
-                parent.Children.RemoveAt(descendentChildIndex);
-            }
-            else
-            {
-                childIndex = GetChildIndex(parent, child);
-            }
-
-            parent.Children.Insert(childIndex, child);
-            child.Parent = parent;
-            _eventsToNodes.Add(child.HistoryEvent, child);
-
-            // Insert into horizontal events!
-            int newHorizontalIndex = GetHorizontalInsertionIndex(parent, childIndex);
-
-            _horizontalNodes.Insert(newHorizontalIndex, child);
-
-            RefreshHorizontalPositions(newHorizontalIndex);
-
-            // Insert vertically!
-            int verticalIndex = GetVerticalIndex(child);
-            _verticalNodes.Insert(verticalIndex, child);
-
-            RefreshVerticalPositions(verticalIndex);
-
-
-            // Raise some kind of event here!
-
-
-            return child;
-        }
-
-
-
-        private bool AddEventToListTree(HistoryEvent historyEvent)
-        {
-            HistoryEvent parentHistoryEvent = historyEvent.Parent;
-            ListTreeNode<HistoryEvent> parentNode = GetNode(parentHistoryEvent);
-            bool wasParentInteresting = parentNode != null;
-            bool isParentInteresting = InterestingEvent(parentHistoryEvent);
-            ListTreeNode<HistoryEvent> node = GetNode(historyEvent);
-
-            if (node != null)
-            {
-                // The node is already in the tree... we shouldn't be trying to add it!
-                throw new Exception("Node was already in the tree.");
-            }
-
-            if (!InterestingEvent(historyEvent))
-            {
-                return false;
-            }
-
-            bool add = true;
-
-            // First, check if the parent's interestingness has changed from false to true.
-            if (!wasParentInteresting && isParentInteresting)
-            {
-                // Need to add the parent!
-
-                // But first, find the child who will share this new parent!
-                ListTreeNode<HistoryEvent> cousinNode = null;
-                HistoryEvent he = parentHistoryEvent;
-                while (true)
-                {
-                    he = he.Children[0];
-                    cousinNode = GetNode(he);
-                    if (cousinNode != null)
-                    {
-                        break;
-                    }
-                }
-
-                parentNode = InsertNewParent(cousinNode, parentHistoryEvent);
-            }
-            else if (!wasParentInteresting && !isParentInteresting)
-            {
-                // Work our way up the tree to find who should be our parent!
-                HistoryEvent he = parentHistoryEvent;
-                while (true)
-                {
-                    ListTreeNode<HistoryEvent> n = GetNode(he);
-                    if (n != null)
-                    {
-                        parentNode = n;
-                        break;
-                    }
-
-                    he = he.Parent;
-                }
-            }
-            else if (wasParentInteresting && !isParentInteresting)
-            {
-                // Replace the parent node with the child!
-                Update(parentHistoryEvent, historyEvent);
-                add = false;
-
-                // Todo: Notify!
-            }
-            else if (wasParentInteresting && isParentInteresting)
-            {
-                // Nothing to do! Just add the new node!
-            }
-
-            if (add && InterestingEvent(historyEvent))
-            {
-                Add(parentNode, historyEvent);
-
-                // Todo: Notify!
-            }
-
-            return true;
-        }
-
-        private NotifyPositionChangedEventArgs<HistoryEvent> UpdateListTree(HistoryChangedEventArgs args)
-        {
-            //if (_listTree == null)
-            //{
-            //    return null;
-            //}
-
-            bool changed = false;
-
-            //lock (_listTree)
-            {
-                switch (args.Action)
-                {
-                    case HistoryChangedAction.Add:
-                        {
-                            changed = AddEventToListTree(args.HistoryEvent);
-                        }
-                        break;
-                    case HistoryChangedAction.UpdateCurrent:
-                        {
-                            ListTreeNode<HistoryEvent> node = GetNode(args.HistoryEvent);
-
-                            changed = Update(node);
-                        }
-                        break;
-                    case HistoryChangedAction.DeleteBranch:
-                        {
-                            ListTreeNode<HistoryEvent> node = GetNode(args.HistoryEvent);
-
-                            RemoveRecursive(node);
-
-                            changed = true;
-                        }
-                        break;
-                    case HistoryChangedAction.DeleteBookmark:
-                        {
-                            ListTreeNode<HistoryEvent> node = GetNode(args.HistoryEvent);
-
-                            RemoveNonRecursive(node);
-
-                            changed = true;
-                        }
-                        break;
-                }
-            }
-
-            if (changed)
-            {
-                NotifyPositionChangedEventArgs<HistoryEvent> changeArgs = new NotifyPositionChangedEventArgs<HistoryEvent>(HorizontalOrdering(), VerticalOrdering());
-
-                return changeArgs;
-            }
-
-            return null;
-        }
-
-        protected override int HorizontalSort(HistoryEvent x, HistoryEvent y)
-        {
-            UInt64 yMax = y.GetMaxDescendentTicks();
-            UInt64 xMax = x.GetMaxDescendentTicks();
-            int result = yMax.CompareTo(xMax);
-
-            if (result != 0)
-            {
-                return result;
-            }
-
-            return y.Id.CompareTo(x.Id);
-        }
-
-        protected override int VerticalSort(HistoryEvent x, HistoryEvent y)
-        {
-            if (_verticalTies.Any())
-            {
-                _verticalTies.Remove(new Tuple<HistoryEvent, HistoryEvent>(x, y));
-                _verticalTies.Remove(new Tuple<HistoryEvent, HistoryEvent>(y, x));
-            }
-
-            if (x.Ticks < y.Ticks)
-            {
-                return -1;
-            }
-            else if (x.Ticks > y.Ticks)
-            {
-                return 1;
-            }
-            else
-            {
-                if (ReferenceEquals(x, y))
-                {
-                    return 0;
-                }
-                else if (x.IsEqualToOrAncestorOf(y))
-                {
-                    return -1;
-                }
-                else if (y.IsEqualToOrAncestorOf(x))
-                {
-                    return 1;
-                }
-            }
-
-            _verticalTies.Add(new Tuple<HistoryEvent, HistoryEvent>(x, y));
-            _verticalTies.Add(new Tuple<HistoryEvent, HistoryEvent>(y, x));
-
-            return x.Id.CompareTo(y.Id);
-        }
-
-
-        //public List<ListTreeNode<HistoryEvent>> HorizontalOrdering
-        //{
-        //    get
-        //    {
-        //        return HorizontalOrdering();
-        //    }
-        //}
-
-        //public List<ListTreeNode<HistoryEvent>> VerticalOrdering
-        //{
-        //    get
-        //    {
-        //        return VerticalOrdering();
-        //    }
-        //}
-
-        public event NotifyPositionChangedEventHandler<HistoryEvent> PositionChanged;
-
-        private History _history;
-        //private ListTree<HistoryEvent> _listTree;
-    }
-
-    public enum NotifyListChangedAction
-    {
-        Added,
-        Moved,
-        Replaced,
-        Removed,
-        Cleared
-    }
-
-    public class NotifyPositionChangedEventArgs<T> : EventArgs
-    {
-        public NotifyPositionChangedEventArgs(List<ListTreeNode<T>> horizontalOrdering, List<ListTreeNode<T>> verticalOrdering)
-        {
-            // Make copies of the orderings... don't use the "live" copies of them!
-            HorizontalOrdering = horizontalOrdering.ToList();
-            VerticalOrdering = verticalOrdering.ToList();
-        }
-
-        public List<ListTreeNode<T>> HorizontalOrdering { get; }
-        public List<ListTreeNode<T>> VerticalOrdering { get; }
-    }
-
-    public delegate void NotifyPositionChangedEventHandler<T>(object sender, NotifyPositionChangedEventArgs<T> e);
-
-
-
 }
